pyproject.toml
[tool.poetry]
name = "legit_io"
version = "0.1.0"
description = ""
authors = ["Z1R343L <jan2705g@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
twitchio = "^2.0.5"
python-dotenv = "^0.19.0"
"ruamel.yaml" = "^0.17.10"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
